#  Switch TransactionQueue from a std::queue to a std::forward_list. 