--- a/src/include/transaction/transaction_defs.h
+++ b/src/include/transaction/transaction_defs.h
@@ -1,7 +1,6 @@
 #pragma once
 
-#include <list>
-#include <queue>
+#include <forward_list>
 namespace terrier::transaction {
 class TransactionContext;
 // Explicitly define the underlying structure of std::queue as std::list since we believe the default (std::deque) may
@@ -10,5 +9,5 @@ class TransactionContext;
 // automatically (it's a linked list). We can afford std::list's pointer-chasing because it's only processed in a
 // background thread (GC). This structure can be replace with something faster if it becomes a measurable performance
 // bottleneck.
-using TransactionQueue = std::queue<transaction::TransactionContext *, std::list<transaction::TransactionContext *>>;
+using TransactionQueue = std::forward_list<transaction::TransactionContext *>;
 }  // namespace terrier::transaction